pyproject.toml
[project]
name = "dbt_gdpr_anonymizer"
version = "0.1.0"
description = "GDPR-compliant dbt anonymization pipeline for French public data"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dbt-core>=1.11.6",
    "dbt-duckdb>=1.10.1",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "pandas>=3.0.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "rich>=14.3.2",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.1",
]


[tool.uv] 
package = true
